scripts/format-compose/pyproject.toml
[project]
name = "format-compose"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ruamel-yaml>=0.18.0",
]
